pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-private"
version = "0.1.0"
description = "Multi-agent MCP framework for Cursor IDE"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "chromadb>=0.4",
    "langfuse>=3.0",
    "sentence-transformers>=2.2",
    "python-dotenv",
    "PyYAML>=6.0",
    "cachetools>=5.3",
    "mcp>=0.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "anthropic>=0.18.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
